worker-service/requirements.txt
pika
minio
google-api-python-client
google-auth
